# docs/openapi.json
{
  "components": {
    "parameters": {
...
      "AttachmentId": {
        "in": "path",
        "name": "attachment_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
          "type": "string"
        }
      },
...
      "NoteId": {
        "in": "path",
        "name": "note_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "note or authorization is not visible"
      },
...
      "Unavailable": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "the service is temporarily unavailable"
      },
...
    },
    "schemas": {
...
      "DiagnosticSeverity": {
        "enum": [
          "error",
          "warning",
          "information",
          "hint"
        ],
        "type": "string"
      },
...
      "NoteDiagnostic": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NoteSourcePosition"
              },
              {
                "type": "null"
              }
            ],
            "description": "本文上の1始まりの表示位置。列はUTF-16 code unit単位で、LSPの既定位置符号化と一致する。"
          },
          "severity": {
            "$ref": "#/components/schemas/DiagnosticSeverity"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Utf8ByteSpan"
              },
              {
                "type": "null"
              }
            ]
          },
          "target": {
            "$ref": "#/components/schemas/NoteValidationTarget"
          }
        },
        "required": [
          "code",
          "severity",
          "target",
          "message"
        ],
        "type": "object"
      },
...
      "NoteSourcePosition": {
        "additionalProperties": false,
        "properties": {
          "column": {
            "format": "uint32",
            "minimum": 1,
            "type": "integer"
          },
          "line": {
            "format": "uint32",
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "line",
          "column"
        ],
        "type": "object"
      },
...
      "NoteValidationTarget": {
        "description": "入力上の問題が、ノート入力のどの部分にあるかを示す位置。\n\nREST、MCP、Web UIで同じ表現を使用する。`field`を判別子とし、`tag`と`acl_entry`は\n対象の添字を伴う。",
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "field": {
                "const": "source",
                "type": "string"
              }
            },
            "required": [
              "field"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "field": {
                "const": "title",
                "type": "string"
              }
            },
            "required": [
              "field"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "field": {
                "const": "body",
                "type": "string"
              }
            },
            "required": [
              "field"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "field": {
                "const": "tag",
                "type": "string"
              },
              "index": {
                "format": "uint",
                "minimum": 0,
                "type": "integer"
              }
            },
            "required": [
              "field",
              "index"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "field": {
                "const": "tags",
                "type": "string"
              }
            },
            "required": [
              "field"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "field": {
                "const": "acl_entry",
                "type": "string"
              },
              "index": {
                "format": "uint",
                "minimum": 0,
                "type": "integer"
              }
            },
            "required": [
              "field",
              "index"
            ],
            "type": "object"
          }
        ]
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/NoteDiagnostic"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ProblemCode": {
        "enum": [
          "authentication_required",
          "authentication_unavailable",
          "csrf_rejected",
          "csrf_required",
          "csrf_invalid",
          "same_origin_required",
          "origin_not_allowed",
          "not_found",
          "forbidden",
          "conflict",
          "retention_expired",
          "invalid_sync_cursor",
          "sync_cursor_expired",
          "precondition_required",
          "invalid_request",
          "patch_rejected",
          "validation_failed",
          "advisories_rejected",
          "render_failed",
          "unavailable"
        ],
        "type": "string"
      },
...
      "Utf8ByteSpan": {
        "additionalProperties": false,
        "properties": {
          "end": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "start": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "unit": {
            "$ref": "#/components/schemas/Utf8ByteUnit"
          }
        },
        "required": [
          "start",
          "end",
          "unit"
        ],
        "type": "object"
      },
      "Utf8ByteUnit": {
        "enum": [
          "utf8_byte"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OAuthAccessToken": {
        "description": "WWW-Authenticateが示すProtected Resource Metadataから取得するOAuth 2.1 access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Marginalis REST API",
    "version": "v3",
    "x-adocweave-package-version": "0.41.0",
    "x-note-profile-version": 18
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/notes/{note_id}/attachments/{attachment_id}/content": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "image bytes"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "summary": "Read one authorized note image"
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/NoteId"
        },
        {
          "$ref": "#/components/parameters/AttachmentId"
        }
      ]
    },
...
  }
}